Add extra test for certificatesService

+++ MyWebsiteBackend.Tests/src/Services/CertificatesServiceTest.cs
@@ -15,16 +15,61 @@ public class CertificatesServiceTests : IDisposable, IClassFixture<TestFixture>
         _context = fixture.Context;
     }
 
+    protected virtual void Dispose(bool disposing)
+    {
+        if (disposing)
+        {
+            _context.Certificates.RemoveRange(_context.Certificates);
+            _context.SaveChanges();
+        }
+    }
+
+    // This code added to correctly implement the disposable pattern.
     public void Dispose()
     {
-        _context.Certificates.RemoveRange(_context.Certificates);
-        _context.SaveChanges();
+        // Do not change this code. Put cleanup code in Dispose(bool disposing) above.
+        Dispose(true);
+        GC.SuppressFinalize(this);
     }
 
     [Fact]
     public async Task GetCertificates_ReturnsAllCertificates()
     {
-        var data = new List<Certificate>
+        var data = GenerateCertificates();
+
+        _context.Certificates.AddRange(data);
+        _context.SaveChanges();
+
+        var service = new CertificatesService(_context);
+
+        // Act
+        var certificates = await service.GetCertificates();
+
+        // Assert
+        Assert.Equal(3, certificates.Count());
+    }
+
+    [Fact]
+    public async Task GetCertificate_ReturnsCertificateById()
+    {
+        var data = GenerateCertificates();
+        var id = data.First().Id;
+
+        _context.Certificates.AddRange(data);
+        _context.SaveChanges();
+
+        var service = new CertificatesService(_context);
+
+        // Act
+        var certificate = await service.GetCertificateById(id);
+
+        // Assert
+        Assert.Equal(id, certificate!.Id);
+    }
+
+    private IEnumerable<Certificate> GenerateCertificates()
+    {
+        return new List<Certificate>
         {
             new Certificate
             {
@@ -48,16 +93,5 @@ public class CertificatesServiceTests : IDisposable, IClassFixture<TestFixture>
                 IssueDate = DateTime.Now
             },
         };
-
-        _context.Certificates.AddRange(data);
-        _context.SaveChanges();
-
-        var service = new CertificatesService(_context);
-
-        // Act
-        var certificates = await service.GetCertificates();
-
-        // Assert
-        Assert.Equal(3, certificates.Count());
     }
 }
